# doc/api.yaml
openapi: 3.0.0  
info:
  title: WasaPhoto APIs
  description: |
   Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
   upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Username:
          type: object
          description: A string that uniquely identifies the user's profile on WasaPhoto.
          title: Username
          properties:
            Username:
              type: string
              minLength: 3
              maxLength: 30
              pattern: '^.*?$'
              example: Eren_yeager
...
    Photo:
      type: object
      title: photo
      description: A photo
      properties:
        PhotoId: 
          type: integer
          description: |
            The unique Id of the Photo in the User PhotoStream.
            Is the same PhotoId in photos/{PhotoId}
          example: 152164
        Likesnumber:
          type: integer
          description: The number of likes of the photo.
          example: 15
        Commentsnumber:
          type: integer
          description: The number of comments under the photo.
          example: 3
        Dateofupload:
          type: string
          format: date-time
          minLength: 20
          maxLength: 30
          pattern: '^.*?$'
          example: 2018-03-20T09:12:28Z
    User:
      type: object
      title: user
      description: The user profile
      properties:
        Username:
          type: string
          description: A string that uniquely identifies the user's profile on WasaPhoto.
          minLength: 3
          maxLength: 30
          pattern: '^.*?$'
          example: Eren_yeager
        Followers:
          type: integer
          description: The number of followers of the user.
          example: 15
        Following: 
          type: integer
          description: The number of following of the user.
          example: 26
        Numberofphotos:
          type: integer
          description: The number of photos that have been posted by the user
        UploadedPhotos:
          type: array
          title: UploadedPhotos
          description: The photos posted by the user in reverse chronological order.
          minItems: 0
          maxItems: 10000000
          items: { $ref: "#/components/schemas/Photo"}
          example: 
            - likesnumber: 2
              commentsnumber: 1
              dateofupload: 2022-03-20T09:12:28Z
            - likesnumber: 12
              commentsnumber: 5
              dateofupload: 2021-11-20T09:42:18Z
  parameters: 
    Username:
      name: Username
      in: path
      required: true
      description: The Username is a string that uniquely identifies a user in WasaPhoto.
      schema:
        type: string
        description: A string of minimum 3 and maximum 30 characters representing the username.
        minLength: 3
        maxLength: 30
        pattern: '^.*?$'
        example: Eren_Yeager
    PhotoId:
      name: PhotoId
      in: path
      required: true
      description: Photo Id
      schema:
        type: integer
        description: An integer that represents the PhotoId.
        readOnly: true
        example: 5557
...
    FollowerId:
      name: FollowerId
      in: path
      required: true
      description: The FollowerId is a string (the username) that uniquely identifies a user in WasaPhoto.
      schema:
        type: string
        minLength: 3
        maxLength: 30
        description:  A string of minimum 3 and maximum 30 characters representing the FollowerId. 
        pattern: '^.*?$'
        example: Founder_Titan_Official
...
  responses:
    BadRequest:
      description: |
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    InternalServerError:
      description: |
        The server encountered an internal error.
        Further info in server logs.      
paths:
...
  /users/:
    get:
      tags:
        - user
      security:
        - bearerAuth: []
      operationId: getUserProfile
      summary: Get the profile of the user you are looking for.
      description: |
        Return the profile of the user searched via username.
        If the user exists,the user profile will be returned in the response.
      parameters:
        - name: username
          in: query
          required: true
          description: The username of the user you are looking for.
          schema:
            description: A string of minimum 3 and maximum 30 characters representing the username.
            type: string
            minLength: 3
            maxLength: 30
            pattern: '^.*?$'

      responses:
        200: 
          description: Request fullfilled.
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400: 
           $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - user
      security:
        - bearerAuth: []
      operationId: addUser
      summary: Adds a user to WasaPhoto.
      description: |
        Adds a user to WasaPhoto,using the unique username specified in the object in the requestBody.
        The User profile will be returned in the response,more specifically, the new profile will have 
        no photos,no followers and no following.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
      responses: 
        201:
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
  /users/{Username}/:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - user
      security:
        - bearerAuth : []
      operationId: getMyStream
      summary: Get the personalized  photostream for a user.
      description: |
        Returns a collection of photos recently uploaded by the other users followed by the user.
        More specifically,a user's photostream consists of recently uploaded photos (in reverse
        chronological order) of the users followed.
      responses:
        "200": 
          description: Request fullfilled.
          content: 
            application/json:
              schema:
                type: array
                description: An array representing the photostream.
                minItems: 0
                maxItems: 10000000
                items: { $ref: "#/components/schemas/Photo" }
        400: 
           $ref: "#/components/responses/BadRequest"
        404:
          description: |
            User not found.
        500:
          $ref: "#/components/responses/InternalServerError"  
    delete:
      tags:
        - user
      security:
        - bearerAuth : []
      operationId: deleteUser
      summary: Delete a user.
      description: |
        Deletes the user with the specified username.
      responses:
        204:
          description: Delete successfully performed.
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: User with specified username not found in the system.
        500:
          $ref: "#/components/responses/InternalServerError"  
...
  /users/{Username}/photos/:
    parameters:
      - $ref: "#/components/parameters/Username"
    post:
      tags:
        - photos
      security:
        - bearerAuth : []
      operationId: uploadPhoto
      summary: Uploads a Photo
      description: |
        Uploads a photo using the object in the request body.
        The server will create a new unique ID, the client can find it
        in the response.
      requestBody: 
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Photo"
      responses: 
        201:
          description: Photo uploaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
  /users/{Username}/photos/{PhotoId}/:
    parameters:
      - $ref: "#/components/parameters/Username"
      - $ref: "#/components/parameters/PhotoId"
    delete:
      tags:
        - photos
      security:
        - bearerAuth : []
      operationId: deletePhoto
      summary: Deletes a photo from the user uploaded photos.
      description: |
        Deletes the photo with the specified id from the system.
      responses:
        204:
          description: Photo deleted successfully.
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: Photo with specified Id not found.
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /users/{Username}/followers/{FollowerId}:
    parameters:
      - $ref: "#/components/parameters/Username"
      - $ref: "#/components/parameters/FollowerId"
      
    delete:
      tags:
       - user
      security:
        - bearerAuth : []
      operationId: unfollowUser
      summary: Unfollows a user.
      description: |
        Deletes the user with the specified username (../{followerId}) from the list of
        followers of a specific user(../users/{username}).
        A user can remove HIS name from the followers list of another user.
      responses:
        204:
          description: Follower removed successfully.
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: Follower with specified username not found in the system.
        500:
          $ref: "#/components/responses/InternalServerError"
...